mai_assistant/app/requirements.txt
langchain==0.0.343
openai==1.3.6
langserve==0.0.36
sse_starlette==1.8.2
redis==5.0.1
python-dotenv==1.0.0